# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HandoffRail",
    "description": "Session-continuity middleware for multi-agent AI workflows",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/hooks": {
      "get": {
        "tags": [
          "hooks"
        ],
        "summary": "List Webhooks",
        "description": "List all webhooks for the authenticated tenant.",
        "operationId": "list_webhooks_api_v1_hooks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebhookResponse"
                  },
                  "type": "array",
                  "title": "Response List Webhooks Api V1 Hooks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hooks"
        ],
        "summary": "Create Webhook",
        "description": "Register a new webhook URL for packet status change notifications.\n\nThe webhook will receive POST requests with HMAC-SHA256 signed payloads\nwhen the subscribed events occur.",
        "operationId": "create_webhook_api_v1_hooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hooks/{webhook_id}": {
      "delete": {
        "tags": [
          "hooks"
        ],
        "summary": "Delete Webhook",
        "description": "Deactivate (soft-delete) a webhook. The webhook is marked as inactive.\n\nInactive webhooks will not receive notifications but are retained for\naudit purposes.",
        "operationId": "delete_webhook_api_v1_hooks__webhook_id__delete",
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Webhook Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hooks/dlq": {
      "get": {
        "tags": [
          "hooks"
        ],
        "summary": "List Dlq",
        "description": "List dead letter queue entries for the authenticated tenant.",
        "operationId": "list_dlq_api_v1_hooks_dlq_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Dlq Api V1 Hooks Dlq Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookCreate": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Webhook callback URL"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events",
            "description": "Event types to subscribe to"
          },
          "secret": {
            "type": "string",
            "maxLength": 256,
            "minLength": 16,
            "title": "Secret",
            "description": "Secret for HMAC-SHA256 signing"
          }
        },
        "type": "object",
        "required": [
          "url",
          "secret"
        ],
        "title": "WebhookCreate",
        "description": "Request body for registering a webhook."
      },
      "WebhookResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "events",
          "tenant_id",
          "active",
          "created_at"
        ],
        "title": "WebhookResponse",
        "description": "Response for a registered webhook."
      }
    }
  }
}